feat: optimizing App router
