+++ components/index.ts
@@ -1,4 +1,3 @@
-export { default as ProjectCard } from './ProjectCard';
 export { default as Navbar } from './Navbar';
 export { default as Footer } from './Footer';
 export { default as ScrollToTop } from './ScrollToTop';
